onboard_bolt/requirements.txt
atlassian-python-api
python-dotenv
ipython
langchain
langchain-openai
slack_bolt
langchain-pinecone